+++ yash-cli/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The shell now supports declaration utilities as defined in POSIX.
 - The `cd` built-in now supports the `-e` option as defined in POSIX.
+- The `trap` built-in now implements the POSIX.1-2024 behavior of showing
+  signal dispositions that are not explicitly set by the user. It also supports
+  the `-p` (`--print`) option.
 - The `-p` option for the `command` built-in now works on Linux.
 
 ### Changed